requirements.txt
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=8.3.0
opencv-python>=4.5.0
albumentations>=1.3.0
timm>=0.6.0
segmentation-models-pytorch>=0.3.0
monai>=1.0.0
nibabel>=4.0.0
SimpleITK>=2.2.0
wandb>=0.13.0
grad-cam>=1.4.0
scipy>=1.7.0
tqdm>=4.62.0
torchmetrics>=0.10.0
efficientnet-pytorch-3d>=1.0.0
medpy>=0.4.0